change int to long in PageNumber

+++ src/Models/QueryFilter.cs
@@ -5,9 +5,9 @@ namespace Behlog.Core.Models;
 /// </summary>
 public class QueryFilter
 {
-    public long PageNumber { get; set; } = 1;
+    public int PageNumber { get; set; } = 1;
     public int PageSize { get; set; } = 10;
-    public long StartIndex => (PageNumber * PageSize) - PageSize;
+    public int StartIndex => (PageNumber * PageSize) - PageSize;
     public string OrderBy { get; set; }
     public bool OrderDesc { get; set; }
